fix(core): apply core 64-bit and thread stability fixes (#242)
* fix(core): resolve 64-bit truncation bug in system timer

* fix(ui): fix pointer truncation in GadgetListBox GetSelection for 64-bit systems

* fix(network): rethrow abi::__forced_unwind in GameSpy threads

* fix(deploy): handle static FFmpeg linkage gracefully in Linux deploy scripts

* fix(deploy): filter DXVK log flood from console output in deploy scripts

+++ Core/Libraries/Source/WWVegas/WWLib/systimer.h
@@ -48,11 +48,11 @@
 #else
 #include <sys/time.h>
 
-inline unsigned long systimerGetMS()
+inline unsigned int systimerGetMS()
 {
 	struct timeval tv;
 	gettimeofday(&tv, nullptr);
-	return (tv.tv_sec * 1000) + (tv.tv_usec / 1000);
+	return (unsigned int)((tv.tv_sec * 1000) + (tv.tv_usec / 1000));
 }
 
 #define TIMEGETTIME systimerGetMS
@@ -75,9 +75,9 @@ class SysTimeClass
 		/*
 		** Get. Use everywhere you would use timeGetTime
 		*/
-		WWINLINE unsigned long Get();
-		WWINLINE unsigned long operator () () {return(Get());}
-		WWINLINE operator unsigned long() {return(Get());}
+		WWINLINE unsigned int Get();
+		WWINLINE unsigned int operator () () {return(Get());}
+		WWINLINE operator unsigned int() {return(Get());}
 
 		/*
 		** Use periodically (like every few days!) to make sure the timer doesn't wrap.
@@ -94,12 +94,12 @@ class SysTimeClass
 		/*
 		** Time we were first called.
 		*/
-		unsigned long StartTime;
+		unsigned int StartTime;
 
 		/*
 		** Time to add after timer wraps.
 		*/
-		unsigned long WrapAdd;
+		unsigned int WrapAdd;
 
 };
 
@@ -120,7 +120,7 @@ extern SysTimeClass SystemTime;
  * HISTORY:                                                                                    *
  *   10/25/2001 1:38PM ST : Created                                                            *
  *=============================================================================================*/
-WWINLINE unsigned long SysTimeClass::Get()
+WWINLINE unsigned int SysTimeClass::Get()
 {
 	/*
 	** This has to be static here since we don't know if we will get called in a global constructor of another object before our
@@ -133,7 +133,7 @@ WWINLINE unsigned long SysTimeClass::Get()
 		is_init = true;
 	}
 
-	unsigned long time = timeGetTime();
+	unsigned int time = (unsigned int)timeGetTime();
 	if (time > StartTime) {
 		return(time - StartTime);
 	}
